Made some variables unsigned

+++ Engine/Components/MeshRendererComponent.cpp
@@ -40,7 +40,7 @@ unsigned ae3d::MeshRendererComponent::New()
 
 Material* ae3d::MeshRendererComponent::GetMaterial( int subMeshIndex )
 {
-    return subMeshIndex < materials.count ? materials[ subMeshIndex ] : nullptr;
+    return subMeshIndex < (int)materials.count ? materials[ subMeshIndex ] : nullptr;
 }
 
 ae3d::MeshRendererComponent* ae3d::MeshRendererComponent::Get( unsigned index )
@@ -226,7 +226,7 @@ void ae3d::MeshRendererComponent::Render( const Matrix44& localToView, const Mat
     }
 }
 
-void ae3d::MeshRendererComponent::SetMaterial( Material* material, int subMeshIndex )
+void ae3d::MeshRendererComponent::SetMaterial( Material* material, unsigned subMeshIndex )
 {
     if (subMeshIndex >= 0 && subMeshIndex < materials.count )
     {
